backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Validation & Settings
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Authentication
google-auth>=2.0.0
python-jose[cryptography]>=3.3.0

# AI
google-genai>=1.0.0

# HTTP Client (testing, external calls)
httpx>=0.27.0
